move parameters to its own file
